requirements.txt
alembic==1.14.0
fastapi==0.115.4
openpyxl==3.1.5
passlib==1.7.4
pillow==11.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
uvicorn==0.32.0

sqlmodel~=0.0.22
PyJWT~=2.10.1